requirements.txt
pandas
numpy
python-pptx
reportlab
statsmodels
